refactor to separate out client side js. Get node info via a call to out local server. Store Voltage API key in dotenv file to avoid adding it to source control
