+++ app.js
@@ -1,6 +1,16 @@
 const express = require("express");
 const app = express();
 
+// Add CORS to allow requests from out client on localhost
+const cors = require("cors");
+const corsOptions = {
+  origin: "*",
+  credentials: true, //access-control-allow-credentials:true
+  optionSuccessStatus: 200,
+};
+
+app.use(cors(corsOptions));
+
 // Use dotenv file for storing backend secrets like API key
 const dotenv = require("dotenv");
 dotenv.config();
@@ -16,46 +26,28 @@ app.get("/", async (req, res) => {
 // endpoint that returns json of your lighning node
 // will be required later on
 app.get("/getNodeInfo", async (req, res) => {
-  getNodeDetails().then((data) => {
+  let data = await getNodeDetails();
+  if (data) {
     return res.send(data);
-  });
+  } else {
+    res.status(500).send("no nodes found, check your API key");
+  }
 });
 
 // call voltage API and request data about your lightning node
 async function getNodeDetails() {
   const endpoint = new URL("https://api.voltage.cloud/node");
-  const API_KEY = process.env.VOLTAGE_API_KEY; // <------- insert your API key
+  const API_KEY = process.env.VOLTAGE_API_KEY;
   const response = await fetch(endpoint, {
     headers: {
       "X-VOLTAGE-AUTH": API_KEY,
     },
   });
   const data = await response.json();
-  return data.nodes[0];
-}
-
-// get details of your node and then display them
-async function displayDetails() {
-  getNodeDetails().then((data) => {
-    const nodeInfo = data;
-    checkForExistingElement(data);
-    for (var key in nodeInfo) {
-      const node = document.createElement("div");
-      node.setAttribute("id", key);
-      var textnode = document.createTextNode(key + ": " + nodeInfo[key]);
-      node.appendChild(textnode);
-      document.body.appendChild(node);
-    }
-  });
-}
-
-// check if elements are already displayed
-// if they are then delete them
-function checkForExistingElement(nodeInfo) {
-  const elements = document.getElementsByTagName("div");
-  for (let i = elements.length - 1; i >= 0; i--) {
-    const div = elements[i];
-    div.parentElement.removeChild(div);
+  if (data.nodes) {
+    return data.nodes[0];
+  } else {
+    return null;
   }
 }
