+++ index.html
@@ -3,12 +3,12 @@
 
 <head>
     <meta charset="UTF-8">
-    <script type="text/javascript" src="app.js"></script>
+    <script type="text/javascript" src="./client.js"></script>
     <title>Node Information Page</title>
 </head>
 
 <body>
-    <button type="button" onclick="displayDetails()">Click to get info</button>
+    <button type="button" onclick="displayNodeDetails()">Click to get info</button>
 </body>
 
 </html>
