+++ package.json
@@ -8,8 +8,10 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
+        "cors": "^2.8.5",
         "dotenv": "^16.0.3",
-        "express": "^4.18.2"
+        "express": "^4.18.2",
+        "process": "^0.11.10"
     },
     "description": ""
 }
